+++ source/game_sa/Enums/ePedRace.h
@@ -1,0 +1,9 @@
+#pragma once
+
+enum ePedRace {
+    RACE_DEFAULT = 0,  //< What the fuck is a default race?
+    RACE_BLACK = 1,    //< 'B'
+    RACE_WHITE = 2,    //< 'W'
+    RACE_ORIENTAL = 3, //< 'O', 'I'
+    RACE_HISPANIC = 4  //< 'H'
+};
