+++ source/game_sa/Population.h
@@ -7,20 +7,13 @@
 #pragma once
 
 #include "eLevelName.h"
+#include "ePedRace.h"
 
 class CLoadedCarGroup;
 class CPed;
 class CDummy;
 class CDummyObject;
 class CZone;
-
-enum ePedRace {
-    RACE_DEFAULT = 0,  //< What the fuck is a default race?
-    RACE_BLACK = 1,    //< 'B'
-    RACE_WHITE = 2,    //< 'W'
-    RACE_ORIENTAL = 3, //< 'O', 'I'
-    RACE_HISPANIC = 4  //< 'H'
-};
 
 enum eUpdatePedCounterState {
     UPDATE_PED_COUNTER_INCREASE = 0,
